insert data ke calculations tabel

+++ database/migrations/2022_01_07_100936_create_calculations_table.php
@@ -17,9 +17,9 @@ class CreateCalculationsTable extends Migration
             $table->increments('id');
             $table->integer('retailer_id')->unsigned();
             $table->foreign('retailer_id')->references('id')->on('retailers')->onDelete('cascade');
-            $table->integer('cfactor');
-            $table->integer('sfactor');
-            $table->integer('total_score');
+            $table->float('cfactor');
+            $table->float('sfactor');
+            $table->float('total_score');
             $table->timestamps();
         });
     }
